requirements.txt
numpy==1.24.3
requests==2.31.0
python-dotenv==1.1.0
pymongo==4.11.3
python-dateutil==2.9.0
